+++ src/utils/errors.ts
@@ -1,8 +1,10 @@
 export class CliError extends Error {
   exitCode: number;
+  code?: string;
 
-  constructor(message: string, exitCode = 1) {
+  constructor(message: string, exitCode = 1, code?: string) {
     super(message);
     this.exitCode = exitCode;
+    this.code = code;
   }
 }